Update PHPUnit configuration
This change enables display of tests that trigger deprecations, errors,
warnings, and notices. The reason for that is that I was seeing that
some of these were being triggered, but I didn't know where or why.
Given that I don't want any of these to occur, I'm making this
configuration change to display them if/when they occur.