Register TwilioHandler as a DI service
This change, as the title states, registers a new service with the DI
container, one that provides a TwilioHandler. To simplify this, and out
of necessity, it uses PHP Dotenv to retrieve the required environment
variables from a dotenv file (.env). These are the user's Twilio account
SID, auth token, and phone number, as well as the application's base URL.