Refactor and rename findUserIDByEmailAddress to return full user details
The reason for this is to avoid creating a new method to retrieve all
user details so that they can be supplied to the call to
TwilioHandler::sendNewNoteNotification in ProcessRequestHandler. To add
another method when one already exists that returns a user's ID seems
like code duplication and a waste of time.