Implement the ability to send an SMS notification after a note is created
This is a small-ish change (bigger than it should be) that uses the
TwilioHandler to send a notification to the user, telling them about the
note being added to their account, once that note has been added on
their account. It also moves the add note functionality into a utility
function to make it easier to maintain.