Add a new function to validate if reference IDs are linked to a user
On writing the README I realised that, despite requiring a reference ID,
it was only superficially checked, but without ever linking said
reference ID to a user. So, the reference ID was effectively only
superficially used. This change starts to correct that oversight by
adding a function in the DatabaseHandler to validate if a reference ID
is linked to a user.