Reduce the number of args to DatabaseHandler::insertAttachment
This change reduces the number of arguments
DatabaseHandler::insertAttachment to two: a note ID and an Attachment.
The reason for doing this is to make the code less verbose and not use
arguments for no reason. In this case, an Argument object was available,
so it made sense to just pass it to the function, instead of passing the
various aspects of the Argument to the function.